backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["adityanandanx <nandanaditya985@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
Flask = "^3.0.3"
langchain-core = "^0.3.15"
langgraph = "^0.2.45"
langchain-google-genai = "^2.0.4"
python-dotenv = "^1.0.1"
langchain-community = "^0.3.5"
mkdocs = "^1.6.1"
mkdocs-material = "^9.5.44"
langchain-openai = "^0.2.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
